UNOMI-467: fix ContextRequest and EventCollectorRequest objects deserialization on GET request with 'payload' parameter, also refactor a bit the structure of the packages to have better visibility on classes and there roles inside the rest bundle
