+++ rest/src/main/java/org/apache/unomi/rest/exception/RuntimeExceptionMapper.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.unomi.rest;
+package org.apache.unomi.rest.exception;
 
 import org.osgi.service.component.annotations.Component;
